requirements.txt
alembic==1.14.1
blinker==1.9.0
click==8.1.3
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.2.0
Jinja2==3.1.2
Mako==1.3.9
MarkupSafe==2.1.1
python-decouple==3.8
SQLAlchemy==2.0.37
typing_extensions==4.12.2
Werkzeug==3.1.3
WTForms==3.0.1
zipp==3.8.1
